chore: Rename RequestParamUtils to RequestParamsValidator [TECH-1603]
